refactor: Moved new-admin page to admin page (with the new sidebar)
