+++ .env
@@ -4,4 +4,4 @@
 # Prisma supports the native connection string format for PostgreSQL, MySQL, SQLite, SQL Server, MongoDB and CockroachDB.
 # See the documentation for all the connection string options: https://pris.ly/d/connection-strings
 
-DATABASE_URL="mongodb+srv://dextroussafe:ZQfPNdpcaPzruuEd@gdsc-csusm.vlxtu.mongodb.net/GDSC-CSUSM?retryWrites=true&w=majority&appName=GDSC-CSUSM"
+DATABASE_URL="mongodb+srv://dsccsusm:hp1NKZyVsjLeBFs2@gdsc-csusm.igsu0.mongodb.net/GDSC-CSUSM?retryWrites=true&w=majority&appName=GDSC-CSUSM"
