+++ app/admin/page.tsx
@@ -1,34 +1,14 @@
-import React from "react";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import EventsTable from "@/components/main/admin/events/events-table";
-import UsersTable from "@/components/main/admin/users/users-table";
-import OverviewCards from "@/components/main/admin/overview/overview-cards";
 import OverviewAnalytics from "@/components/main/admin/overview/overview-analytics";
+import OverviewCards from "@/components/main/admin/overview/overview-cards";
+import Container from "@/components/ui/container";
+import React from "react";
 
 const Dashboard = () => {
   return (
-    <main className="w-full flex-center-col py-[6rem]">
-      {/* PAGE CONTENT */}
-      <div className="custom-max-width">
-        <Tabs defaultValue="overview" className="w-full">
-          <TabsList className="custom-box-shadow">
-            <TabsTrigger value="overview">Overview</TabsTrigger>
-            <TabsTrigger value="users">Users</TabsTrigger>
-            <TabsTrigger value="events">Events</TabsTrigger>
-          </TabsList>
-          <TabsContent value="users">
-            <UsersTable />
-          </TabsContent>
-          <TabsContent value="overview" className="flex flex-col gap-6">
-            <OverviewCards />
-            <OverviewAnalytics />
-          </TabsContent>
-          <TabsContent value="events">
-            <EventsTable />
-          </TabsContent>
-        </Tabs>
-      </div>
-    </main>
+    <Container className="custom-max-width flex flex-col justify-center -mt-20 gap-8">
+      <OverviewCards />
+      <OverviewAnalytics />
+    </Container>
   );
 };
