+++ components/main/admin/create-event/success-message.tsx
@@ -32,7 +32,7 @@ const SuccessMessage = () => {
       initial="hidden"
       animate="visible"
     >
-      <Link href="/new-admin/events" className="absolute top-0 left-0 flex gap-2 text-sm text-priamry hover:underline">
+      <Link href="/admin/events" className="absolute top-0 left-0 flex gap-2 text-sm text-priamry hover:underline">
         <ChevronLeft size={20}/> Back to Events
       </Link>
       <Image
@@ -52,7 +52,7 @@ const SuccessMessage = () => {
       <div className="flex items-center mt-6">
         <div className="relative after:pointer-events-none after:absolute after:inset-px after:rounded-[11px] after:shadow-highlight after:shadow-white/10 focus-within:after:shadow-[#77f6aa] after:transition">
           <Button
-            onClick={() => router.push("/new-admin/events")}
+            onClick={() => router.push("/admin/events")}
             className="relative bg-blue border group border-border hover:bg-blue/70 rounded-xl text-white hover:text-white"
           >
             <FaArrowRight className="mr-2 h-4 w-4 group-hover:translate-x-1 transition-transform" />
